+++ lib/bones_api_adapter_postgre.dart
@@ -1,4 +1,4 @@
 /// Bones_API DB Adapter for PostgreSQL.
-library bones_api_adapter.postgre;
+library bones_api_adapter.postgres;
 
-export 'src/bones_api_entity_adapter_postgre.dart';
+export 'src/bones_api_entity_adapter_postgres.dart';