+++ test/bones_api_test.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/1.0.20
+// BUILDER: reflection_factory/1.0.21
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -32,7 +32,7 @@ class MyInfoModule$reflection extends ClassReflection<MyInfoModule> {
   Version get languageVersion => Version.parse('2.13.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.0.20');
+  Version get reflectionFactoryVersion => Version.parse('1.0.21');
 
   @override
   MyInfoModule$reflection withObject([MyInfoModule? obj]) =>